Update ROSELCDd.py
For some unknown reasons, the sec2ms(sec) function was not working that well as the seek value was getting tampered by 10 secs from time to time....I looked for a new function that would not need to modify the seek rraw result in milliseconds (prevent the use of player.seek[:3] ) I found the following page  https://stackoverflow.com/questions/35989666/convert-milliseconds-to-hours-min-and-seconds-python . And leveraging the code to create a new ms2mns function that translate milliseconds in hours, minutes and seconds.  I do not expose hours in this code.